+++ packages/mobile/src/backup/BackupIntroduction.tsx
@@ -1,35 +1,30 @@
-import Button, { BtnTypes } from '@celo/react-components/components/Button'
-import colors from '@celo/react-components/styles/colors'
-import { fontStyles } from '@celo/react-components/styles/fonts'
+import Button from '@celo/react-components/components/Button.v2'
+import TextButton from '@celo/react-components/components/TextButton.v2'
+import { default as colors, default as colorsV2 } from '@celo/react-components/styles/colors.v2'
+import fontStyles from '@celo/react-components/styles/fonts.v2'
+import { Sizes } from '@celo/react-components/styles/styles.v2'
 import * as React from 'react'
-import { WithTranslation } from 'react-i18next'
+import { useTranslation, WithTranslation } from 'react-i18next'
 import { ScrollView, StyleSheet, Text, View } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import { connect } from 'react-redux'
-import { setBackupDelayed } from 'src/account/actions'
 import CeloAnalytics from 'src/analytics/CeloAnalytics'
 import { CustomEventNames } from 'src/analytics/constants'
 import componentWithAnalytics from 'src/analytics/wrapper'
 import { enterBackupFlow, exitBackupFlow } from 'src/app/actions'
+import DelayButton from 'src/backup/DelayButton'
 import { Namespaces, withTranslation } from 'src/i18n'
-import BackupIntroIcon from 'src/icons/BackupIntroIcon'
-import BackupKeyIcon from 'src/icons/BackupKeyIcon'
-import SafeguardsIcon from 'src/icons/SafeguardsIcon'
+import Logo from 'src/icons/Logo.v2'
 import { headerWithBackButton } from 'src/navigator/Headers'
-import { navigate, navigateBack, navigateProtected } from 'src/navigator/NavigationService'
+import { navigateProtected } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { RootState } from 'src/redux/reducers'
-import { isBackupTooLate } from 'src/redux/selectors'
 
 interface StateProps {
   backupCompleted: boolean
-  socialBackupCompleted: boolean
-  backupTooLate: boolean
-  backupDelayedTime: number
 }
 
 interface DispatchProps {
-  setBackupDelayed: typeof setBackupDelayed
   enterBackupFlow: typeof enterBackupFlow
   exitBackupFlow: typeof exitBackupFlow
 }
@@ -39,15 +34,13 @@ type Props = WithTranslation & StateProps & DispatchProps
 const mapStateToProps = (state: RootState): StateProps => {
   return {
     backupCompleted: state.account.backupCompleted,
-    socialBackupCompleted: state.account.socialBackupCompleted,
-    backupTooLate: isBackupTooLate(state),
-    backupDelayedTime: state.account.backupDelayedTime,
   }
 }
 
 class BackupIntroduction extends React.Component<Props> {
   static navigationOptions = () => ({
     ...headerWithBackButton,
+    headerRight: <DelayButton />,
   })
 
   componentDidMount() {
@@ -58,169 +51,101 @@ class BackupIntroduction extends React.Component<Props> {
     this.props.exitBackupFlow()
   }
 
-  onPressViewBackupKey = () => {
-    CeloAnalytics.track(CustomEventNames.view_backup_phrase)
-    navigateProtected(Screens.BackupPhrase)
-  }
-
   onPressBackup = () => {
     CeloAnalytics.track(CustomEventNames.set_backup_phrase)
     navigateProtected(Screens.BackupPhrase)
   }
 
-  onPressSetupSocialBackup = () => {
-    CeloAnalytics.track(CustomEventNames.set_social_backup)
-    navigate(Screens.BackupSocialIntro)
-  }
-
-  onPressViewSocialBackup = () => {
-    CeloAnalytics.track(CustomEventNames.view_social_backup)
-    navigateProtected(Screens.BackupSocial)
-  }
-
-  onPressDelay = () => {
-    this.props.setBackupDelayed()
-    CeloAnalytics.track(CustomEventNames.delay_backup)
-    navigateBack()
-  }
-
   render() {
-    const {
-      t,
-      backupDelayedTime,
-      backupTooLate,
-      backupCompleted,
-      socialBackupCompleted,
-    } = this.props
+    const { backupCompleted } = this.props
     return (
       <SafeAreaView style={styles.container}>
-        <ScrollView contentContainerStyle={styles.scrollContainer}>
-          {!backupCompleted && (
-            <>
-              <BackupIntroIcon style={styles.logo} width={180} height={138} />
-              <Text style={styles.h1}>{t('setUpBackupKey')}</Text>
-              <Text style={styles.body}>{t('backupKeyIntro.0')}</Text>
-              <Text style={[styles.body, fontStyles.bold]}>{t('backupKeyIntro.1')}</Text>
-              <Text style={styles.body}>{t('backupKeyIntro.7')}</Text>
-            </>
-          )}
-          {backupCompleted && !socialBackupCompleted && (
-            <>
-              <SafeguardsIcon style={styles.logo} width={210} height={90} />
-              <Text style={styles.h1}>{t('setUpSocialBackup')}</Text>
-              <Text style={styles.body}>
-                {t('backupKeyIntro.2')}
-                <Text style={[styles.body, fontStyles.bold]}>{t('backupKeyIntro.3')}</Text>
-              </Text>
-              <Text style={styles.body}>{t('backupKeyIntro.4')}</Text>
-            </>
-          )}
-          {backupCompleted && socialBackupCompleted && (
-            <>
-              <BackupKeyIcon style={styles.logo} width={170} height={125} />
-              <Text style={styles.h1}>{t('backupComplete.header')}</Text>
-              <Text style={styles.body}>
-                {t('backupKeyIntro.2')}
-                <Text style={[styles.body, fontStyles.bold]}>{t('backupKeyIntro.3')}</Text>
-              </Text>
-              <Text style={styles.body}>
-                {t('backupKeyIntro.5')}
-                <Text style={[styles.body, fontStyles.bold]}>{t('backupKeyIntro.6')}</Text>
-              </Text>
-            </>
-          )}
-        </ScrollView>
-        <View>
-          {!backupCompleted && (
-            <>
-              <Button
-                onPress={this.onPressBackup}
-                text={t('getYourKey')}
-                standard={false}
-                type={BtnTypes.PRIMARY}
-              />
-              {backupTooLate && !backupDelayedTime && (
-                <Button
-                  onPress={this.onPressDelay}
-                  text={t('delayBackup')}
-                  standard={false}
-                  type={BtnTypes.SECONDARY}
-                />
-              )}
-            </>
-          )}
-
-          {backupCompleted && !socialBackupCompleted && (
-            <>
-              <Button
-                onPress={this.onPressSetupSocialBackup}
-                text={t('setUpSocialBackup')}
-                standard={false}
-                type={BtnTypes.PRIMARY}
-              />
-              <Button
-                onPress={this.onPressViewBackupKey}
-                text={t('viewBackupKey')}
-                standard={false}
-                type={BtnTypes.SECONDARY}
-              />
-            </>
-          )}
-
-          {backupCompleted && socialBackupCompleted && (
-            <>
-              <Button
-                onPress={this.onPressBackup}
-                text={t('viewBackupKey')}
-                standard={false}
-                type={BtnTypes.SECONDARY}
-              />
-              <Button
-                onPress={this.onPressViewSocialBackup}
-                text={t('viewSafeguards')}
-                standard={false}
-                type={BtnTypes.SECONDARY}
-              />
-            </>
-          )}
-        </View>
+        {backupCompleted ? (
+          <AccountKeyPostSetup />
+        ) : (
+          <AccountKeyIntro onPrimaryPress={this.onPressBackup} />
+        )}
       </SafeAreaView>
     )
   }
 }
 
+interface AccountKeyStartProps {
+  onPrimaryPress: () => void
+}
+
+function AccountKeyIntro({ onPrimaryPress }: AccountKeyStartProps) {
+  const { t } = useTranslation(Namespaces.accountKeyFlow)
+  return (
+    <ScrollView contentContainerStyle={styles.introContainer}>
+      <Logo height={32} />
+      <Text style={styles.h1}>{t('introTitle')}</Text>
+      <Text style={styles.body}>{t('introBody')}</Text>
+      <Button text={t('introPrimaryAction')} onPress={onPrimaryPress} />
+    </ScrollView>
+  )
+}
+// TODO in next PR get Real mneumonic
+function AccountKeyPostSetup() {
+  const { t } = useTranslation(Namespaces.accountKeyFlow)
+  return (
+    <ScrollView contentContainerStyle={styles.postSetupContainer}>
+      <Text style={fontStyles.h2}>{t('introTitle')}</Text>
+      <View style={styles.keyArea}>
+        <Text style={fontStyles.large}>
+          horse leopard dog monkey shark tiger lemur whale squid wolf squirrel mouse lion elephant
+          cat shrimp bear penguin deer turtle fox zebra goat giraffe
+        </Text>
+      </View>
+      <Text style={styles.postSetupBody}>{t('postSetupBody')}</Text>
+      <View style={styles.postSetupCTA}>
+        <TextButton>{t('postSetupCTA')}</TextButton>
+      </View>
+    </ScrollView>
+  )
+}
+
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: colors.background,
-    justifyContent: 'space-between',
+    backgroundColor: colors.light,
   },
-  scrollContainer: {
+  introContainer: {
     flex: 1,
-    paddingHorizontal: 30,
-    paddingBottom: 30,
+    paddingHorizontal: Sizes.THICK,
     justifyContent: 'center',
   },
-  logo: {
-    alignSelf: 'center',
-    marginBottom: 30,
+  postSetupContainer: {
+    paddingTop: Sizes.THICK,
+    flex: 1,
+    paddingHorizontal: Sizes.REG,
   },
   h1: {
     ...fontStyles.h1,
+    paddingBottom: Sizes.REG,
+    paddingTop: Sizes.REG,
   },
   body: {
-    ...fontStyles.body,
-    textAlign: 'center',
-    paddingBottom: 15,
+    ...fontStyles.large,
+    paddingBottom: Sizes.REG,
   },
-  loader: {
-    marginBottom: 20,
+  keyArea: {
+    padding: Sizes.REG,
+    backgroundColor: colorsV2.brownFaint,
+    marginTop: Sizes.REG,
+  },
+  postSetupBody: {
+    ...fontStyles.regular,
+    marginVertical: Sizes.REG,
+  },
+  postSetupCTA: {
+    alignSelf: 'center',
+    marginTop: Sizes.REG,
   },
 })
 
 export default componentWithAnalytics(
   connect<StateProps, DispatchProps, {}, RootState>(mapStateToProps, {
-    setBackupDelayed,
     enterBackupFlow,
     exitBackupFlow,
   })(withTranslation(Namespaces.backupKeyFlow6)(BackupIntroduction))
