+++ packages/mobile/src/backup/BackupPhrase.tsx
@@ -1,9 +1,9 @@
-import Button, { BtnTypes } from '@celo/react-components/components/Button'
-import Switch from '@celo/react-components/components/Switch'
-import colors from '@celo/react-components/styles/colors'
-import { fontStyles } from '@celo/react-components/styles/fonts'
+import Button, { BtnSizes, BtnTypes } from '@celo/react-components/components/Button.v2'
+import Switch from '@celo/react-components/components/Switch.v2'
+import colors from '@celo/react-components/styles/colors.v2'
+import fontStyles from '@celo/react-components/styles/fonts.v2'
 import * as React from 'react'
-import { WithTranslation } from 'react-i18next'
+import { useTranslation, WithTranslation } from 'react-i18next'
 import { ScrollView, StyleSheet, Text, View } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import { connect } from 'react-redux'
@@ -17,10 +17,12 @@ import BackupPhraseContainer, {
   BackupPhraseType,
 } from 'src/backup/BackupPhraseContainer'
 import { getStoredMnemonic } from 'src/backup/utils'
+import CancelButton from 'src/components/CancelButton.v2'
 import { Namespaces, withTranslation } from 'src/i18n'
-import { headerWithBackButton } from 'src/navigator/Headers'
+import { headerWithCancelButton } from 'src/navigator/Headers'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
+import TopBarButton from 'src/navigator/TopBarButton.v2'
 import { RootState } from 'src/redux/reducers'
 import Logger from 'src/utils/Logger'
 
@@ -49,9 +51,10 @@ const mapStateToProps = (state: RootState): StateProps => {
 class BackupPhrase extends React.Component<Props, State> {
   // TODO(Rossy): Show modal when cancelling if backup flow incomplete
   static navigationOptions = () => ({
-    ...headerWithBackButton,
+    ...headerWithCancelButton,
+    headerLeft: <CancelButton color={colors.gray4} />,
+    headerRight: <HeaderRight />,
   })
-
   state = {
     mnemonic: '',
     isConfirmChecked: false,
@@ -100,23 +103,15 @@ class BackupPhrase extends React.Component<Props, State> {
     return (
       <SafeAreaView style={styles.container}>
         <ScrollView contentContainerStyle={styles.scrollContainer}>
-          <View>
-            <Text style={fontStyles.h1}>{t('yourBackupKey')}</Text>
-            <Text style={styles.body}>{t('backupKeySummary')}</Text>
-            <BackupPhraseContainer
-              value={mnemonic}
-              showCopy={true}
-              mode={BackupPhraseContainerMode.READONLY}
-              type={BackupPhraseType.BACKUP_KEY}
-            />
-            <Text style={styles.tipText}>
-              <Text style={[styles.tipText, fontStyles.bold]}>{t('global:warning')}</Text>
-              {t('securityTip')}
-            </Text>
-          </View>
+          <BackupPhraseContainer
+            value={mnemonic}
+            mode={BackupPhraseContainerMode.READONLY}
+            type={BackupPhraseType.BACKUP_KEY}
+          />
+          <Text style={styles.body}>{t('backupKeySummary')}</Text>
         </ScrollView>
         {!backupCompleted && (
-          <View>
+          <>
             <View style={styles.confirmationSwitchContainer}>
               <Switch value={isConfirmChecked} onValueChange={this.onPressConfirmSwitch} />
               <Text style={styles.confirmationSwitchLabel}>{t('savedConfirmation')}</Text>
@@ -125,47 +120,45 @@ class BackupPhrase extends React.Component<Props, State> {
               disabled={!isConfirmChecked}
               onPress={this.onPressContinue}
               text={t('global:continue')}
-              standard={false}
-              type={BtnTypes.PRIMARY}
+              size={BtnSizes.FULL}
+              type={BtnTypes.SECONDARY}
             />
-          </View>
+          </>
         )}
       </SafeAreaView>
     )
   }
 }
 
+function HeaderRight() {
+  const { t } = useTranslation(Namespaces.backupKeyFlow6)
+  return <TopBarButton>{t('moreInfo')}</TopBarButton>
+}
+
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: colors.background,
+    backgroundColor: colors.light,
     justifyContent: 'space-between',
+    padding: 16,
   },
   scrollContainer: {
     flexGrow: 1,
-    paddingHorizontal: 20,
-    paddingBottom: 20,
+    paddingBottom: 16,
   },
   body: {
-    ...fontStyles.body,
-    marginBottom: 20,
-  },
-  tipText: {
-    ...fontStyles.bodySmall,
-    color: colors.darkSecondary,
-    marginTop: 25,
-    marginHorizontal: 3,
+    ...fontStyles.regular,
+    marginTop: 16,
   },
   confirmationSwitchContainer: {
-    paddingHorizontal: 20,
-    paddingVertical: 20,
+    paddingVertical: 16,
     flexDirection: 'row',
+    alignItems: 'center',
   },
   confirmationSwitchLabel: {
-    ...fontStyles.body,
-    ...fontStyles.semiBold,
-    paddingTop: 3,
-    paddingLeft: 10,
+    flex: 1,
+    ...fontStyles.regular,
+    paddingLeft: 8,
   },
 })
