+++ packages/mobile/src/components/CancelButton.v2.tsx
@@ -1,3 +1,4 @@
+import { colorsEnum } from '@celo/react-components/styles/colors.v2'
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
 import CeloAnalytics from 'src/analytics/CeloAnalytics'
@@ -7,11 +8,12 @@ import { navigateBack } from 'src/navigator/NavigationService'
 import TopBarButton from 'src/navigator/TopBarButton.v2'
 
 interface Props {
+  color?: colorsEnum
   eventName?: CustomEventNames
   onCancel?: () => void
 }
 
-export default function CancelButton({ eventName, onCancel }: Props) {
+export default function CancelButton({ eventName, onCancel, color }: Props) {
   const onPressCancel = React.useCallback(() => {
     if (eventName) {
       CeloAnalytics.track(eventName)
@@ -27,7 +29,7 @@ export default function CancelButton({ eventName, onCancel }: Props) {
   const { t } = useTranslation(Namespaces.global)
 
   return (
-    <TopBarButton testID="CancelButton" onPress={onPressCancel}>
+    <TopBarButton testID="CancelButton" onPress={onPressCancel} style={{ color }}>
       {t('cancel')}
     </TopBarButton>
   )
